engine: extract esm/video_output_type.js
